Initial Commit
The important task of controlling hate speech on internet platforms begins with the task of identifying a piece of hate speech. Due to dearth of labeled data, we present two ways in which we can classify texts as hate speech with limited number of labeled samples and an unlabeled dataset. We also perform hate speech classification using a model trained with only positive and unlabeled samples.

Co-Authored-By: 4dityarao <40599446+4dityarao@users.noreply.github.com>